requirements.txt
asgiref==3.6.0
Django==4.1.7
django-tinymce==3.5.0
python-decouple==3.7
sqlparse==0.4.3
